modified:   ansible.cfg 	renamed:    go.mod -> apiansible/go.mod 	renamed:    go.sum -> apiansible/go.sum 	new file:   apiansible/main.go 	new file:   hosts 	new file:   local.yml 	deleted:    main.go 	renamed:    add-nfs-playbook.yml -> playbooks/add-nfs-playbook.yml 	renamed:    install-nfs-playbook.yml -> playbooks/install-nfs-playbook.yml 	renamed:    remove-nfs-playbook.yml -> playbooks/remove-nfs-playbook.yml 	renamed:    uninstall-nfs-playbook.yml -> playbooks/uninstall-nfs-playbook.yml 	deleted:    set-ip-server-playbook.yml
