+++ ansible.cfg
@@ -12,5 +12,4 @@
 # o
 [default]
 inventory=/etc/ansible/hosts
-host_key_checking = False
