+++ hosts
@@ -1,0 +1,2 @@
+[local]
+127.0.0.1
