--- set-ip-server-playbook.yml
@@ -1,32 +1,0 @@
----
-- name: set ip
-  hosts: new-host
-  remote_user: ansible-user
-  become: true
-  gather_facts: false
-
-  tasks:
-    - name: Write the new ec2 instance host key to known hosts
-      connection: local
-      shell: "ssh-keyscan -H {{ inventory_hostname }} >> ~/.ssh/known_hosts"
-
-
-    - name: ipchange
-      nmcli:
-        conn_name: ens33
-        ifname: ens33
-        type: ethernet
-        ip4: "{{ ip }}"
-        gw4: 10.4.187.1
-        state: present
-    
-    - name: restart service
-      service:
-        name: restart
-        enabled: false
-        state: started
-    
-
-    
-
-
